refactor(SaveGame): DBConnection을 인터페이스로 분리
- H2DBConnection은 구현체로 분리
