+++ src/test/java/repository/GameDaoTest.java
@@ -19,7 +19,7 @@ class GameDaoTest {
 
     @BeforeEach
     void setUp() {
-        DBConnection dbConnection = new DBConnection(TEST_URL);
+        DBConnection dbConnection = new H2DBConnection(TEST_URL);
         gameDao = new GameDao(dbConnection);
 
         try (Connection connection = dbConnection.getConnection();
